refactor: 사용자 프로필 fetch 및 비밀번호 변경 로직을 서비스 함수로 분리
