+++ src/containers/mypage/ProfileSectionContainer.jsx
@@ -7,7 +7,7 @@ import { updateProfile } from "firebase/auth";
 
 import { auth, db } from "services/firebase";
 import { logout } from "services/auth/auth";
-import { changeUserPassword } from "services/userService";
+import { changeUserPassword } from "services/user/changeUserPassword";
 
 import { setUser, clearUser } from "store/userSlice";
 import { useUserStats } from "hooks/mypage/useUserStats";
